chore: Remove unused components
- Removed unused component, `Linha` and `Principal`.
- Renamed `Barra` and `BarraTotal` files to improve developer experience.
